Allow subclasses to work without an explicit opt-in.
This removes the requirement that every job class `include Sidekiq::Symbols`,
even if a parent class already did it. Say you have this structure:

  class ApplicationJob
    include Sidekiq::Worker
    include Sidekiq::Symbols

    # ...
  end

  class FooJob < ApplicationJob
    def perform(x: 1)
      # ...
    end
  end

Enqueueing a job for `FooJob` would _not_ allow symbol usage in the
existing, previous implementation, because `ApplicationJob` is the class
that `Sidekiq::Symbols` was prepended in front of.

  ApplicationJob
        |
  Sidekiq::Symbols
        |
      FooJob

So `FooJob` (and every other subclass of `ApplicationJob`) had to
explicitly duplicate the `include Sidekiq::Symbols` line unless it
called `super` or something similar.

This helps address GitHub issue 2.